// Web app/csi-mkd-premarital-app-FE/src/api/api-main-app/models/class-feedback-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClassFeedbackDetailDto } from '../models/class-feedback-detail-dto';
export interface ClassFeedbackDto {
  email: string;
  feedbacks: {
[key: string]: ClassFeedbackDetailDto;
};
  name: string;
  premaritalRegistrationId: string;
}