// Web app/csi-mkd-premarital-app-FE/src/api/api-main-app/models/class-feedback-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClassFeedbackDetailDto } from '../models/class-feedback-detail-dto';
export interface ClassFeedbackResponseDto {
  createdAt?: string;
  email?: string | null;
  feedbacks?: ({
[key: string]: ClassFeedbackDetailDto;
}) | null;
  id?: number;
  name?: string | null;
  premaritalRegistrationId?: string;
  updatedAt?: string;
}